refactor(core): rename internal core modules for clarity
