+++ package.json
@@ -66,7 +66,9 @@
     "test:examples": "vitest run examples --coverage",
     "types": "svelte-check",
     "types:legacy": "svelte-check --tsconfig tsconfig.legacy.json",
-    "build": "tsc -p tsconfig.build.json && cp src/component-types.d.ts types",
+    "build": "npm-run-all build:*",
+    "build:tsc": "tsc -p tsconfig.build.json",
+    "build:copy-dts": "cp src/core/types.d.ts types/core",
     "contributors:add": "all-contributors add",
     "contributors:generate": "all-contributors generate",
     "preview-release": "./scripts/preview-release",
