+++ src/core/index.js
@@ -5,15 +5,15 @@
  * Will switch to legacy, class-based mounting logic
  * if it looks like we're in a Svelte <= 4 environment.
  */
-import * as LegacyCore from './legacy.js'
-import * as ModernCore from './modern.svelte.js'
-import { createValidateOptions } from './validate-options.js'
+import * as MountLegacy from './mount-legacy.js'
+import * as MountModern from './mount-modern.svelte.js'
+import { createValidateOptions } from './prepare.js'
 
 const { mount, unmount, updateProps, allowedOptions } =
-  ModernCore.IS_MODERN_SVELTE ? ModernCore : LegacyCore
+  MountModern.IS_MODERN_SVELTE ? MountModern : MountLegacy
 
 /** Validate component options. */
 const validateOptions = createValidateOptions(allowedOptions)
 
 export { mount, unmount, updateProps, validateOptions }
-export { UnknownSvelteOptionsError } from './validate-options.js'
+export { UnknownSvelteOptionsError } from './prepare.js'
