+++ src/core/mount-legacy.js
