+++ src/core/mount-modern.svelte.js
